Added ability to import from Excel (partially) TBD: Dates & IP Addresses
Currently doing it manually with xlrd which reads cell by cell, and handles the IP addresses & dates in an odd way that will have to be modified so for now I left those out. Added a page to allow the user to upload the excel file and submit to import the data. Also modified some frontend to make the table with populated data display properly. Changed some max lengths of charfields in the applicationserver model as well.