All fields successfully importing from excel, changed drives int->dec
Dates are now handled correctly, as well as a few other quirks due to Excel storing all numbers as floats. Modified drive fields to be decimal fields instead of integer to properly store the necessary values.

TODO: Properly validate excel sheet to provide useful error messages to the user when attempting to import invalid data. Provide instructions on the import page on the expected format of the excel sheet