Fixed Accidental Server Creation when Updating a Server
Caused new servers to be created instead of the old one being updated when the user opts to edit a server. This was caused by the get_or_create method used to ensure that duplicate servers are not entered into the database, and calling that same method when editing a server. Of course when editing, the goal is not to create a new server so that logic does not work. Wrote a new function to handle the case of editing a server, and left the function with the get_or_create function as-is for the case of importing from excel or submitting the new server form.