Improved Template Consistency & Style, Fixed Import Bug
Excel Import logic was missing the last row of the sheet, had to add one to the loop ending condition in order to account for that. Also added excel import sample file to show required formatting to end user and make importing from excel more convenient. Still need to fix the import page, the centered form can block the "Download Sample Template" button if the window is resized to be small enough