Added Confirmation before Import of New Entries
New page after import submit which displays new entries that will be added and prompts user to confirm the import. Required new field in the table "visible" (t/f) which defines whether an entry is valid and should be displayed to the user. Also added required login for views that were missing it.