Added logistical fields to server model, fixed Notes field in edit form
Now the ApplicationServer model has published_date, published_by, edited_date, and edited_by fields. Those fields are displayed as appropriate in the application server detail page. Additionally, an oversight was made with the notes field in the edit server form page, as the frontend of that field had not been modified as it was for the create server form page.