Added all missing columns to table export, changed color based on hwdate
Columns were added to the datatable to make it more complete a while ago, but the export definitions did not reflect the additional columns. Now they do. The color of each entry can also now be yellow or red depending on the proximity of the next hardware support date. If the date is within the next month, the table entry will be yellow. If it has already passed, the table entry will be red.