Refined Login
Made display of the login more consistent, got rid of nbcu authentication backend which is not functional yet and only causes timeouts at the moment. Will add that authentication backend back in later, pretty sure all it needs is to be included in the allowed_hosts list and it will work.