Added all_fields to FilterProfile
It is a new field that, when filled while creating a filter profile, queries all columns in the table to check for the value stored in it. So if the value is "Anvato", any application server with a field value which contains "Anvato" or "anvato" will be returned by the query.