Improved Additional IP Hover Functionality
Hover is now animated, table is now the same color as the main table, and there are no longer empty tables when table entries don't have additional IPs, there is now a check to see if there are any additional IPs. If not, "None" is displayed. If so, the "View IPs" link is displayed.